changed (unsigned) to (NSUInteger)
git-svn-id: svn://orca.physics.unc.edu/Orca@11739 fd86f90f-2c0a-422a-b972-84106d8db31b

+++ Source/Main Dialogs/CommandCenter/Scripting/NodeTree.h
@@ -47,8 +47,8 @@
 - (int)				operatorTag;
 - (void)			setOperatorTag:(int)anOperator;
 - (NSMutableArray*) operands;
-- (unsigned)		count;
-- (id)				objectAtIndex:(unsigned)index;
+- (NSUInteger)		count;
+- (id)				objectAtIndex:(NSUInteger)index;
 - (void)			setOperands:(NSMutableArray*)anArray;
 - (void)			addOperand:(id)anOperand;
 @end
